Use structs to apply boundary conditions to a problem
This will allow us to have more complex boundary conditions which can do
some preprocessing when initiated.

It also allows us to have boundary conditions that keep an internal
state, which may be useful for buffer allocations which we can use to
improve performance or (among others) to approximate derivatives.