Rename equilibria folder to velocity_distribution_function
This will allow us to further generalize VDFs including the addition of
VDFs for incompressible flows, using cumulant based VDFs etc.
To do so we will first have to make some changes to the Quadrature
structures so that they can include information about the equilibrium
distribution that should be used.
TODO: look into using traits for the equilibrium functions